Add files via upload
N houses are built in a single row numbered from 0 to N−1. The first house has a house number 0, the second house has a house number 1 and so on, every house pays some rent at the end of the month. Find out the house number of the house paying the minimum rent.
Note: All house rents are unique.